% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCenter.R
\name{findCenter}
\alias{findCenter}
\title{findCenter}
\usage{
findCenter(x, y, lim)
}
\arguments{
\item{x}{x component of distribution}

\item{y}{y component of ditribution}

\item{lim}{a vector of bin limits}
}
\value{
a vector of central values
}
\description{
finds the mode of the distribution within each bin
}
