% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLim.R
\name{findLim}
\alias{findLim}
\title{findLim}
\usage{
findLim(x, y, range = 1L)
}
\arguments{
\item{x}{x component of distribution}

\item{y}{y component of ditribution}

\item{range}{the number of sequential x values to use to detect a bin limit (units of 0.01 bases)}
}
\value{
a vector of bin limits
}
\description{
finds the bin limits based on the estimated distribution of fragment lengths
}
