% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStat.R
\name{findStat}
\alias{findStat}
\title{findStat}
\usage{
findStat(frag, lim, FUN)
}
\arguments{
\item{frag}{vector of raw fragment lengths}

\item{lim}{a vector of bin limits}

\item{FUN}{a function to be applied the fragments within each bin}
}
\value{
a vector of doubbles
}
\description{
returns descriptive statistics about the raw fragment leanghts in each bin
}
