% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDensityPlot.R
\name{ggDensityPlot}
\alias{ggDensityPlot}
\title{ggDensityPlot}
\usage{
ggDensityPlot(frag, x, y, center, lim, title = NULL)
}
\arguments{
\item{frag}{a vector of microsatilite fragment length}

\item{x}{the x component of the density plot}

\item{y}{the y component of the density plot}

\item{center}{a vector of central values of each bin}

\item{lim}{a vector of bin limits}

\item{title}{a title}
}
\value{
a list containing results of bandwidth estimation, KDE, bin limits, central values, mean and frequency.
}
\description{
produces a density plot of fragment lengths alternately colour coded by bin.
}
