% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineLength.R
\name{lineLength}
\alias{lineLength}
\title{lineLength}
\usage{
lineLength(x, y)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}
}
\value{
the length of the line
}
\description{
finds the length of a line drawn between a series of sequential points described by x and y coordinates
}
