% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDatabase.R
\name{makeDatabase}
\alias{makeDatabase}
\title{makeDatabase}
\usage{
makeDatabase(df, ploidy)
}
\arguments{
\item{df}{a data frame like object containing raw SSR fragment length data}

\item{ploidy}{the number of alleles per locus per indiviual.}
}
\value{
a complex named list containing an element for each locus in the input data
and a sub-element for the output of each of SSRr's core functions
}
\description{
A function that builds a repository of named lists from a tidy input database for storing the outputs of other MicroBinR functions.
}
