% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{sat_assess}
\alias{sat_assess}
\title{sat_assess}
\usage{
sat_assess(repo, h = NULL, range = NULL, K = NULL)
}
\arguments{
\item{repo}{a database containing data for each loci (produced by sat_makeDatabase)}

\item{range}{a vector containing ranges for each loci passed to sat_binStats if (null defaults to 20)}

\item{sgma}{a vector containing sigma values for each loci passed to sat_score (if null defaults to 0.4)}
}
\value{
a database updated with the ouputs of sat_score, sat_binStats and sat_plot
}
\description{
runs sat_score, sat_binStats and sat_plot for all loci for which info$complete = "FALSE". The product of each function is saved in a data repository, upon completion the complete placeholder is changes to "TRUE"
}
