% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{sat_assign}
\alias{sat_assign}
\title{sat_assign}
\usage{
sat_assign(repo)
}
\arguments{
\item{repo}{a repository initialised by sat_retrive and updated by sat_linerise, sat_score and sat_bin}
}
\value{
a complex named list containing an element for each locus in the input data
and a sub-element for the output of each of SSRr's core functions
}
\description{
sat_assign updates the sat repository with binned fragment lengths useing the initial data and bin limets produced by sat_bin.
}
