% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{sat_bin}
\alias{sat_bin}
\title{sat_bin}
\usage{
sat_bin(frag_len, limits, center)
}
\arguments{
\item{repo}{a database containing data for each loci (produced by sat_makeDatabase)}

\item{sgma}{a vector containing sigma values for each loci passed to sat_score (if null defaults to 0.4)}

\item{range}{a vector containing ranges for each loci passed to sat_binStats if (null defaults to 20)}
}
\value{
vector of binned fragment lengths
}
\description{
Takes a vector of fragment lengths, lower limits and central values and sorts values into bins
}
