% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binStats.R
\name{sat_binStats}
\alias{sat_binStats}
\title{sat_binStats}
\usage{
sat_binStats(x, z, y, range = 20)
}
\arguments{
\item{x}{raw fragment lengths}

\item{z}{all possible fragment lengths}

\item{y}{approximation of "closeness" of each z value to all x values}

\item{range}{the number of consecutively riseing / falling y values to take into account when finding bin limits}
}
\value{
a tibble containing basic statistics about each allele.
}
\description{
sat_binStats assigns x values to bins and produces a tibble of useful statisitcs for each allele
}
