% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCenter.R
\name{sat_findCenter}
\alias{sat_findCenter}
\title{sat_findCenter}
\usage{
sat_findCenter(z, y, limits)
}
\arguments{
\item{z}{a vector of possible fragment lengths}

\item{y}{a vector of desity of fragment lengths, must be same length as z}

\item{limits}{a vector of limits that can be used to partition z into bins.}
}
\value{
a central z value for each bin.
}
\description{
sat_findCenter finds the z with max y value for each bin defined by limits
}
