% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLimits.R
\name{sat_findLimits}
\alias{sat_findLimits}
\title{sat_findLimits}
\usage{
sat_findLimits(z, y, range = 20)
}
\arguments{
\item{z}{a vector of possible fragment lengths}

\item{y}{a vector of desity of fragment lengths, must be same length as z}

\item{range}{the number of z values consecutively riseing or and falling in either direction.}
}
\value{
A list of z values
}
\description{
sat_findLimits finds z values where "range" zvalues in either direction increase and decrease consecutively
}
