% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStat.R
\name{sat_findStat}
\alias{sat_findStat}
\title{sat_findStat}
\usage{
sat_findStat(x, limits, FUN = length)
}
\arguments{
\item{x}{a vector of raw fragment lengths}

\item{limits}{a vector of bin limits (range of limits must encompass range of x)}

\item{FUN}{the function to be applied to each bin (must produce single numeric for each bin)}
}
\value{
a numeric vector same length as number of bins (length(limits) - 1)
}
\description{
sat_findStat sorts a vector of fragment lengths into bins and applies a function to each bin seporately.
}
