% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{sat_getData}
\alias{sat_getData}
\title{sat_getData}
\usage{
sat_getData(df, locus_num, ploidy)
}
\arguments{
\item{data}{a dataframe like object where the first column is a unique identifier for each genotype and subseqent columns contain raw fragment lengths found for that locus.}
}
\value{

}
\description{
sat_getData returns the ID and fragment length columns for a locus from rectangular input data
}
