% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sat_plot}
\alias{sat_plot}
\title{sat_plot}
\usage{
sat_plot(x, z, y, center, lim.upper, title = NULL)
}
\arguments{
\item{z}{a vector of possible fragment lengths}

\item{y}{a vector of desity of fragment lengths, must be same length as z}

\item{center}{a vector representing the center of each bin}

\item{lim.upper}{a vector representing the upper limit of each bin}
}
\value{

}
\description{
sat_plot plots all possible fragment lengths (z) against the "closness" of all possible fragment lengths to observed raw fragment lengths (y)
}
