% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{sat_score}
\alias{sat_score}
\title{sat_score}
\usage{
sat_score(x, h = 0.4, resolution = 0.01, tail.length = 1, K = NULL)
}
\arguments{
\item{x}{a vector of raw fragment lengths}

\item{resolution}{the smallest detectable difference between raw fragment lengths}

\item{tail.length}{= the size of the buffer regions placed either side of raw fragment lengths}

\item{sigma}{assumed standard deviation of error for each raw fragment length}

\item{e}{an approximation of euilers constant}

\item{FUN}{an expression that is passed to purrr::map_dbl which is used to calculate y values}
}
\value{
a complex named list containing an element for each locus in the input data
and a sub-element for the output of each of SSRr's core functions
}
\description{
sat_score takes a vector of microsatilite raw fragment lengths and summs gausian distributions for
 each input produceing a combined distribution of the whole locus
}
