% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{sat_tidy}
\alias{sat_tidy}
\title{sat_tidy}
\usage{
sat_tidy(data)
}
\arguments{
\item{data}{a dataframe like object where the first column is a unique identifier for each genotype and subseqent columns contain raw fragment lengths found for that locus.}
}
\value{

}
\description{
sat_tidy takes each locus from a the sat database, and puts it into tidy format. An additional duplicate column is added indicating whether raw fragment lengths are identical to fragment lengths befor them for a given locus and individual.
}
