% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy.R
\name{sat_untidy}
\alias{sat_untidy}
\title{sat_untidy}
\usage{
sat_untidy(id, frag_len, allele)
}
\arguments{
\item{id}{each individuals unique identification}

\item{frag_len}{vector of fragment lengths}

\item{allele}{vector indicating which allele the fragment length represents}
}
\value{
a tibble containing a column for ID and a seporate column for each allele
}
\description{
sat_untidy spreads fragment length data over multiple columns useing allele number as a key. opposite of sat_untidy.
}
