% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(frag, h)
}
\arguments{
\item{frag}{a vector of fragment lengths}

\item{h}{bandwidth}
}
\value{
x and y values for an unweighted disribution
}
\description{
Calculates a kernel denstiy estimation without wighting the ouput by n or bandwidth. The resulting distribution is calculated at a uniform reolution of 0.01 fwith fixed tails of 1 - min(x) to 1 + max(x).
}
