% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(h, frag, FUN)
}
\arguments{
\item{h}{bandwidth}

\item{frag}{a vector of fragment lengths}

\item{FUN}{an error function}
}
\value{
a doubble produced by the error function
}
\description{
Produces a unwighted KDE estimation useing score for a given bandwidth and then applies an error function to the resulting diistribution.
}
