% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.locus.R
\name{bin.locus}
\alias{bin.locus}
\title{bin.locus}
\usage{
bin.locus(frag)
}
\arguments{
\item{frag}{a vector of microsatilite fragment lengths}
}
\value{
a list containing results of bandwidth estimation, KDE, bin limits, central values, mean and frequency.
}
\description{
Bins microsatilite fragment lengths useing a clustering operation based on kernel density estimation.
}
